fix(server): eliminate N+1 workflow queries in list_tasks
GET /tasks was issuing one DB query per task that had an issue or PR
number attached (O(N) sequential queries). With a large backlog this
added 400-1000 ms latency on every dashboard refresh.

Fix: bulk-load all workflows with a single workflow_store.list() call
before the loop, build two in-memory HashMaps keyed by
(project_id, repo, issue_number) and (project_id, repo, pr_number),
then do O(1) map lookups per task — zero additional DB queries.

list() already returns rows ORDER BY updated_at DESC, so
HashMap::entry().or_insert_with() naturally retains the most recent
workflow for each key.

Extract the github_webhook handler and its helpers into a new
webhook_routes.rs submodule to keep misc_routes.rs under the 800-line
project limit.

Closes #937
