+++ crates/harness-server/src/http/misc_routes.rs
@@ -7,10 +7,9 @@ use axum::{
 };
 use harness_protocol::methods::RpcRequest;
 use serde_json::json;
-use std::path::{Path as StdPath, PathBuf};
 use std::sync::Arc;
 
-use super::{state::AppState, task_routes};
+use super::state::AppState;
 use crate::{router, task_runner};
 
 pub(crate) async fn health_check(State(state): State<Arc<AppState>>) -> Json<serde_json::Value> {
@@ -178,303 +177,33 @@ pub(crate) async fn handle_rpc(
     }
 }
 
-fn configured_github_webhook_project_root(
-    github: Option<&harness_core::config::intake::GitHubIntakeConfig>,
-    default_root: &StdPath,
-    repo: &str,
-) -> Option<PathBuf> {
-    github?
-        .effective_repos()
-        .into_iter()
-        .find(|repo_cfg| repo_cfg.repo == repo)
-        .map(|repo_cfg| {
-            repo_cfg
-                .project_root
-                .map(PathBuf::from)
-                .unwrap_or_else(|| default_root.to_path_buf())
-        })
-}
-
-enum GitHubWebhookProjectRootError {
-    RepoNotConfigured(String),
-    RegistryLookup(String),
-}
-
-fn github_webhook_project_root_error_response(
-    error: GitHubWebhookProjectRootError,
-) -> (StatusCode, Json<serde_json::Value>) {
-    match error {
-        // Treat unknown repositories as ignored so GitHub does not retry
-        // an event for a repo this harness instance is not configured to
-        // serve. Registry failures remain internal errors.
-        GitHubWebhookProjectRootError::RepoNotConfigured(reason) => (
-            StatusCode::OK,
-            Json(json!({ "status": "ignored", "reason": reason })),
-        ),
-        GitHubWebhookProjectRootError::RegistryLookup(error) => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(json!({ "error": error })),
-        ),
-    }
-}
-
-async fn resolve_github_webhook_project_root(
-    state: &Arc<AppState>,
-    repo: &str,
-) -> Result<PathBuf, GitHubWebhookProjectRootError> {
-    if let Some(project_root) = configured_github_webhook_project_root(
-        state.core.server.config.intake.github.as_ref(),
-        &state.core.project_root,
-        repo,
-    ) {
-        return Ok(project_root);
-    }
-
-    if let Some(registry) = state.core.project_registry.as_deref() {
-        if let Some(project) = registry.get(repo).await.map_err(|error| {
-            GitHubWebhookProjectRootError::RegistryLookup(format!(
-                "project registry lookup failed: {error}"
-            ))
-        })? {
-            return Ok(project.root);
-        }
-        if let Some(project) = registry.get_by_name(repo).await.map_err(|error| {
-            GitHubWebhookProjectRootError::RegistryLookup(format!(
-                "project registry lookup failed: {error}"
-            ))
-        })? {
-            return Ok(project.root);
-        }
-    }
-
-    Err(GitHubWebhookProjectRootError::RepoNotConfigured(format!(
-        "webhook repository '{repo}' is not configured in intake.github and was not found in the project registry"
-    )))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{
-        configured_github_webhook_project_root, github_webhook_project_root_error_response,
-        GitHubWebhookProjectRootError,
-    };
-    use axum::http::StatusCode;
-    use harness_core::config::intake::{GitHubIntakeConfig, GitHubRepoConfig};
-    use std::path::PathBuf;
-
-    #[test]
-    fn multi_repo_github_webhook_uses_repo_specific_project_root_override() {
-        let default_root = PathBuf::from("/srv/repo-a");
-        let github = GitHubIntakeConfig {
-            enabled: true,
-            repos: vec![
-                GitHubRepoConfig {
-                    repo: "org/repo-a".to_string(),
-                    label: "harness".to_string(),
-                    project_root: None,
-                },
-                GitHubRepoConfig {
-                    repo: "org/repo-b".to_string(),
-                    label: "harness".to_string(),
-                    project_root: Some("/srv/repo-b".to_string()),
-                },
-            ],
-            ..Default::default()
-        };
-
-        let resolved =
-            configured_github_webhook_project_root(Some(&github), &default_root, "org/repo-b");
-
-        assert_eq!(resolved, Some(PathBuf::from("/srv/repo-b")));
-    }
-
-    #[test]
-    fn configured_github_repo_without_override_falls_back_to_default_project_root() {
-        let default_root = PathBuf::from("/srv/repo-a");
-        let github = GitHubIntakeConfig {
-            enabled: true,
-            repos: vec![GitHubRepoConfig {
-                repo: "org/repo-a".to_string(),
-                label: "harness".to_string(),
-                project_root: None,
-            }],
-            ..Default::default()
-        };
-
-        let resolved =
-            configured_github_webhook_project_root(Some(&github), &default_root, "org/repo-a");
-
-        assert_eq!(resolved, Some(default_root));
-    }
-
-    #[test]
-    fn unconfigured_github_repo_has_no_configured_project_root() {
-        let default_root = PathBuf::from("/srv/repo-a");
-        let github = GitHubIntakeConfig {
-            enabled: true,
-            repos: vec![GitHubRepoConfig {
-                repo: "org/repo-a".to_string(),
-                label: "harness".to_string(),
-                project_root: None,
-            }],
-            ..Default::default()
-        };
-
-        let resolved =
-            configured_github_webhook_project_root(Some(&github), &default_root, "org/repo-b");
-
-        assert_eq!(resolved, None);
-    }
-
-    #[test]
-    fn unconfigured_github_repo_returns_ignored_response() {
-        let (status, body) = github_webhook_project_root_error_response(
-            GitHubWebhookProjectRootError::RepoNotConfigured(
-                "webhook repository 'org/repo-b' is not configured".to_string(),
-            ),
-        );
-
-        assert_eq!(status, StatusCode::OK);
-        assert_eq!(body.0["status"], "ignored");
-        assert!(body.0["reason"]
-            .as_str()
-            .unwrap_or_default()
-            .contains("not configured"));
-    }
-
-    #[test]
-    fn registry_lookup_failures_return_internal_server_error() {
-        let (status, body) = github_webhook_project_root_error_response(
-            GitHubWebhookProjectRootError::RegistryLookup(
-                "project registry lookup failed: boom".to_string(),
-            ),
-        );
-
-        assert_eq!(status, StatusCode::INTERNAL_SERVER_ERROR);
-        assert_eq!(body.0["error"], "project registry lookup failed: boom");
-    }
-}
-
-pub(crate) async fn github_webhook(
-    State(state): State<Arc<AppState>>,
-    headers: HeaderMap,
-    body: Bytes,
-) -> (StatusCode, Json<serde_json::Value>) {
-    let secret = match state
-        .core
-        .server
-        .config
-        .server
-        .github_webhook_secret
-        .as_deref()
-    {
-        Some("") => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(json!({"error": "invalid server.github_webhook_secret configuration"})),
-            )
-        }
-        Some(secret) => secret,
-        None => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(json!({"error": "missing server.github_webhook_secret configuration"})),
-            )
-        }
-    };
-    let signature = match headers
-        .get("x-hub-signature-256")
-        .and_then(|value| value.to_str().ok())
-    {
-        Some(signature) => signature,
-        None => {
-            return (
-                StatusCode::UNAUTHORIZED,
-                Json(json!({"error": "missing header x-hub-signature-256"})),
-            )
-        }
-    };
-    if !crate::webhook::verify_github_signature(secret, signature, body.as_ref()) {
-        return (
-            StatusCode::UNAUTHORIZED,
-            Json(json!({"error": "invalid webhook signature"})),
-        );
-    }
-
-    let event = match headers
-        .get("x-github-event")
-        .and_then(|value| value.to_str().ok())
-    {
-        Some(event) => event,
-        None => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(json!({"error": "missing header x-github-event"})),
-            )
-        }
-    };
-    if !crate::webhook::is_valid_github_event_name(event) {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(json!({"error": "invalid header x-github-event"})),
-        );
-    }
-
-    let (request, reason) =
-        match crate::webhook::parse_github_webhook_task_request(event, body.as_ref()) {
-            Ok(parsed) => parsed,
-            Err(error) => return (StatusCode::BAD_REQUEST, Json(json!({ "error": error }))),
-        };
-
-    let Some(mut req) = request else {
-        return (
-            StatusCode::OK,
-            Json(json!({
-                "status": "ignored",
-                "reason": reason,
-            })),
-        );
-    };
-
-    if req.project.is_none() {
-        req.project = Some(match req.repo.as_deref() {
-            Some(repo) => match resolve_github_webhook_project_root(&state, repo).await {
-                Ok(project_root) => project_root,
-                Err(error) => return github_webhook_project_root_error_response(error),
-            },
-            None => state.core.project_root.clone(),
-        });
-    }
-
-    match task_routes::enqueue_task(&state, req).await {
-        Ok(task_id) => (
-            StatusCode::ACCEPTED,
-            Json(json!({
-                "status": "accepted",
-                "reason": reason,
-                "task_id": task_id.0,
-            })),
-        ),
-        Err(crate::services::execution::EnqueueTaskError::BadRequest(error)) => {
-            (StatusCode::BAD_REQUEST, Json(json!({ "error": error })))
-        }
-        Err(crate::services::execution::EnqueueTaskError::Internal(error)) => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(json!({ "error": error })),
-        ),
-        Err(crate::services::execution::EnqueueTaskError::MaintenanceWindow {
-            retry_after_secs,
-        }) => (
-            StatusCode::SERVICE_UNAVAILABLE,
-            Json(json!({ "error": "maintenance_window", "retry_after": retry_after_secs })),
-        ),
-    }
-}
-
 pub(crate) async fn list_tasks(State(state): State<Arc<AppState>>) -> Response {
     match state.core.tasks.list_all_summaries_with_terminal().await {
         Ok(mut summaries) => {
             if let Some(workflow_store) = state.core.issue_workflow_store.as_ref() {
+                // Bulk-load all workflows once to avoid N+1 queries.
+                // list() returns rows ORDER BY updated_at DESC, so the first
+                // entry for each key is the most recent one.
+                let all_workflows = workflow_store.list().await.unwrap_or_default();
+                let mut issue_map: std::collections::HashMap<
+                    (String, Option<String>, u64),
+                    harness_workflow::issue_lifecycle::IssueWorkflowInstance,
+                > = std::collections::HashMap::new();
+                let mut pr_map: std::collections::HashMap<
+                    (String, Option<String>, u64),
+                    harness_workflow::issue_lifecycle::IssueWorkflowInstance,
+                > = std::collections::HashMap::new();
+                for wf in all_workflows {
+                    let issue_key = (wf.project_id.clone(), wf.repo.clone(), wf.issue_number);
+                    let pr_key = wf
+                        .pr_number
+                        .map(|pr| (wf.project_id.clone(), wf.repo.clone(), pr));
+                    issue_map.entry(issue_key).or_insert_with(|| wf.clone());
+                    if let Some(key) = pr_key {
+                        pr_map.entry(key).or_insert(wf);
+                    }
+                }
+
                 for summary in &mut summaries {
                     let Some(project_id) = summary.project.as_deref() else {
                         continue;
@@ -496,16 +225,14 @@ pub(crate) async fn list_tasks(State(state): State<Arc<AppState>>) -> Response {
                                 .and_then(crate::http::parse_pr_num_from_url)
                         });
                     summary.workflow = match (by_issue, by_pr) {
-                        (Some(issue), _) => workflow_store
-                            .get_by_issue(project_id, summary.repo.as_deref(), issue)
-                            .await
-                            .ok()
-                            .flatten(),
-                        (None, Some(pr)) => workflow_store
-                            .get_by_pr(project_id, summary.repo.as_deref(), pr)
-                            .await
-                            .ok()
-                            .flatten(),
+                        (Some(issue), _) => {
+                            let key = (project_id.to_string(), summary.repo.clone(), issue);
+                            issue_map.get(&key).cloned()
+                        }
+                        (None, Some(pr)) => {
+                            let key = (project_id.to_string(), summary.repo.clone(), pr);
+                            pr_map.get(&key).cloned()
+                        }
                         (None, None) => None,
                     };
                 }
